feat: 참가자에 딜러와 플레이어가 있도록 하라
