+++ README.md
@@ -46,3 +46,16 @@
 - [x] 게임 종료
     - [x] 최종 점수를 계산한다.
     - [x] 결과를 출력한다.
+
+## 🚀 3단계 - 블랙잭(딜러)
+딜러와 플레이어 중 카드의 합이 21 또는 21에 가장 가까운 숫자를 가지는 쪽이 이기는 게임이다.
+
+### 기능 요구사항
+- 딜러는 처음에 받은 2장의 합계가 16이하이면 반드시 1장의 카드를 추가로 받아야 하고, 17점 이상이면 추가로 받을 수 없다.
+- 딜러가 21을 초과하면 그 시점까지 남아 있던 플레이어들은 가지고 있는 패에 상관 없이 승리한다.
+- 게임을 완료한 후 각 플레이어별로 승패를 출력한다.
+
+### 구현할 기능 정의
+- [x] 딜러와 플레이어를 참가자 공통분모로 묶기 
+- [ ] 딜러 초기 카드 분배
+- [ ] 승패 판단
