+++ src/main/kotlin/blackjack/entity/Dealer.kt
@@ -1,0 +1,3 @@
+package blackjack.entity
+
+class Dealer : Participant("딜러")
