+++ src/main/kotlin/blackjack/entity/Player.kt
@@ -1,19 +1,3 @@
 package blackjack.entity
 
-class Player(
-    val name: String,
-) {
-    val hand: Hand = Hand()
-
-    fun receiveCard(card: Card) {
-        hand.addCard(card)
-    }
-
-    fun calculateScore(): Int {
-        return hand.calculateScore()
-    }
-
-    fun isBusted(): Boolean {
-        return hand.isBusted()
-    }
-}
+class Player(name: String) : Participant(name)
